fix(request): return correct remote_addr using trusted_proxies...
The code had a mistake `if ip in trusted_proxies` rather than `if ip in proxy`
which caused that line to always return False meaning the first ip address would
always fall through to the `return str(ip)` line. If the line is fixed, it would
return `None` as it would fall out of the loop and the implicit `return None`
would occur. This logic does not match the logic referenced in the StackOverflow
link.

New, Current Logic:
- go through each address in reverse starting with remote_addr
- Return the first untrusted address, i.e. not contained in any trusted proxy
  networks
- If all addresses are trusted, return remote_addr